requirements.txt
pyaudio==0.2.14
gtts==2.5.4
python-dotenv==1.1.1
requests==2.32.5
pygame==2.6.1
faster-whisper==1.2.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0